requirements.txt
asgiref==3.2.7
autopep8==1.5.4
boto3==1.14.40
botocore==1.17.40
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
defusedxml==0.7.0rc1
dj-database-url==0.5.0
Django==3.0.7
django-bootstrap-modal-forms==2.0.0
django-bootstrap3==12.1.0
django-braces==1.14.0
django-rest-multiple-models==2.1.3
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
drfpasswordless==1.5.6
gunicorn==20.0.4
idna==2.9
isort==5.6.4
jmespath==0.10.0
Pillow==7.1.2
psycopg2==2.8.5
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
stripe==2.48.0
toml==0.10.2
urllib3==1.25.9
